back-end/requirements.txt
Flask
flask-restful
tensorflow
numpy
Pillow
flask-cors